Build serialized tries from sorted keys
`create_serialized_trie` builds a tree of heap-allocated nodes, each
with 256 child pointers, before serializing it breadth-first. For the
CSV reader's default NA values this takes about 15 us on the host, and
every read builds three tries (NA, true and false values).

The builder also keeps node indexes, and the serialized trie its
children offsets, in 16-bit integers. In a trie of more than 32767
nodes, a node can be more than 32767 nodes before its first child, and
the children of such a node were silently lost: the keys below it never
matched. For example, of the 40000 NA values "k00000" to "k39999",
"k39999" did not match, in both the CSV and the JSON reader.

Serialize the trie directly from the distinct keys in sorted order. The
keys below a node are then consecutive, with the key that ends at the
node first, and the others grouped by their next character in ascending
order: each group is a child. A breadth-first walk over (node, key
range) pairs emits the same nodes, in the same order and with the same
offsets, as the tree walk did. `std::string_view` compares characters as
unsigned char, which is the order of the children that
`serialized_trie_contains` relies on for non-ASCII keys.

Store the children offsets in 32-bit integers, so that tries of any
practical size are correct; a trie too large even for that throws
std::overflow_error instead of losing keys. The length of the longest
key now always fits in the root, so lookups no longer handle a root
without it. Nodes grow from 4 to 8 bytes: in the same kernel
measurements as the previous commit, the decode and type detection
kernels show no consistent change.

The output was compared with the previous builder on 30000 random sets
of up to 24 keys (with duplicates, the empty key, and every byte value
but the terminating '\n'): the nodes are identical in every case.
Building the default NA trie takes 0.6 us instead of 15.2 us, and the
default true or false trie 0.13 us instead of 2.9 us. The new CSV,
JSON and TRIE_TEST tests of a trie of 40000 keys, and the TRIE_TEST test
of a key of 40000 characters, fail without this change.

+++ cpp/src/io/utilities/trie.hpp
@@ -34,14 +34,11 @@ static constexpr char trie_terminating_character = '\n';
  * the last character of a word (i.e. `is_leaf` is true).
  *
  * The first node is the root, which matches the empty key. Its children are the nodes that follow
- * it, so its `children_offset` instead holds the length of the longest key, or -1 if that does not
- * fit, which lets lookups reject longer keys without searching the trie.
- *
- * Node indexes and offsets are stored as `int16_t`, so `children_offset` is only meaningful for
- * tries of at most `std::numeric_limits<int16_t>::max()` nodes.
+ * it, so its `children_offset` instead holds the length of the longest key, which lets lookups
+ * reject longer keys without searching the trie.
  */
 struct serial_trie_node {
-  int16_t children_offset{-1};
+  int32_t children_offset{-1};
   char character{trie_terminating_character};
   bool is_leaf{false};
   explicit serial_trie_node(char c, bool leaf = false) noexcept : character(c), is_leaf(leaf) {}
